[IMP] Renamed web_tree_image_tooltip
