+++ web_tree_image_tooltip/__init__.py
@@ -1,0 +1,1 @@
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
