+++ web_tree_image_tooltip/i18n/web_tree_image.pot
