+++ web_tree_image_tooltip/readme/CONTRIBUTORS.rst
@@ -1,0 +1,6 @@
+* Stefan Rijnhart
+* Leonardo Donelli <donelli@webmonks.it>
+* Jay Vora <jay.vora@serpentcs.com>
+* Meet Dholakia <m.dholakia.serpentcs@gmail.com>
+* Nikul Chaudhary <nikul.chaudhary.serpentcs@gmail.com>
+
