+++ web_tree_image_tooltip/readme/CREDITS.rst
@@ -1,0 +1,2 @@
+* Serpent Consulting Services Pvt. Ltd. <support@serpentcs.com>
+
