+++ web_tree_image_tooltip/readme/DESCRIPTION.rst
@@ -1,0 +1,3 @@
+This module defines a images and icons in tree view via Tooltip.
+Additionally, Default width with 30px.
+
