+++ web_tree_image_tooltip/readme/USAGE.rst
@@ -1,0 +1,2 @@
+Mouse Hover in tree view image that time Tooltip effect.
+
