+++ web_tree_image_tooltip/static/src/scss/common.scss
